refactor: api 파일 폴더 정리
