+++ src/api/store/fetchRandomStoreList.tsx
