+++ src/api/store/sendStoreDemandDeleteRequest.tsx
