+++ src/components/pages/CategoryDetailPage/CategoryDetailPage.tsx
@@ -14,8 +14,8 @@ import { PATHNAME } from "constants/routes";
 
 import { campusContext } from "context/CampusContextProvider";
 
-import fetchStoreList from "api/fetchStoreList";
 import getNextPageParam from "api/getNextPageParam";
+import fetchStoreList from "api/store/fetchStoreList";
 
 import Chip from "components/common/Chip/Chip";
 import ErrorImage from "components/common/ErrorImage/ErrorImage";
